Changed OSfrog scoring balance #107
Also added +Rush bonus to the boss kills which was missing before.
On Rush, knowing which boss wave has actually been cleared is a
liability, but I tried to give points for any frog killed past a full
wave clear. This also applies to non-rush games, where the remainder of
modulo 30 would still get added to the final score in the end